Fixed a bug reported on Bioconductor on Thu, 10 Apr 2014, titled "GOseq with ce6". https://stat.ethz.ch/pipermail/bioconductor/attachments/20140410/1c226686/attachment.pl
For C. Elegans, getgo was picking two packages from .ORG due to "grep" matching of the "ce"
string. This caused getgo to fail. I've change "grep" to "match" to avoid this problem, however
the "genome" variable strings are now case sensitive.

The bug reporter also pointed out that getgo does not work for pig, which was caused by the
"Pig" name in .ORG instead of the name given in supportedGenome. This is fixed now.



git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/branches/RELEASE_2_14/madman/Rpacks/goseq@88862 bc3139a8-67e5-0310-9ffc-ced21a209358